task 6 (مخبص)

+++ src/modules/auth/logout.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, HttpStatus } from '@nestjs/common';
+import { Injectable, HttpStatus, HttpException } from '@nestjs/common';
 import { CacheService } from 'core/lib/cache/cache.service';
 import { User } from 'modules/users/entities/user.entity';
 import { UsersService } from 'modules/users/users.service';
@@ -12,12 +12,15 @@ export class LogoutService {
   ) {}
 
   async logUserOut(userID: string): Promise<ResponseFromServiceI<User>> {
-    const userToLogOut = (await this.userService.findOne(userID)).data;
+    const userToLogout = await this.userService.findOneByID(userID);
+    if (!userToLogout)
+      throw new HttpException('user not found', HttpStatus.NOT_FOUND);
     await this.cacheService.deleteField(userID + '', 'accessToken');
+
     return {
       message: 'auth.success.logout',
       httpStatus: HttpStatus.OK,
-      data: userToLogOut,
+      data: userToLogout,
     };
   }
 }
